Add handdrawn 3D examples and two helper scripts
`stack_paths.py` concatenates an n-dimensional path and an m-dimensional path into an (n+m)-dimensional path with as many points as the first path, doing interpolation for the second path.

`transpose_dims.py` takes a path, transposes its dimensions (but can also remove and duplicate dimensions) and writes the resulting path in a new file.